pyproject.toml
[project]
name = "research-platform"
version = "0.1.0"
description = "Multi-Agent Graduate-Level Research Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "google-generativeai>=0.8.0",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.2.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "click>=8.1.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "pyyaml>=6.0.0",
    "websockets>=12.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "plotly>=5.17.0",
    "networkx>=3.2.0",
    "wordcloud>=1.9.0",
    "python-docx>=1.1.0",
    "fpdf2>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.0",
    "pytest-xdist>=3.6.0",
    "pytest-docker>=3.1.0",
    "pytest-benchmark>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-redis>=4.6.0",
    "types-passlib>=1.7.0",
    "pytest-httpx>=0.33.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "ipython>=8.29.0",
    "ipdb>=0.13.0",
    "pre-commit>=4.0.0",
    "testcontainers[postgres,redis]>=4.9.0",
    "playwright>=1.49.0",
    "locust>=2.34.0",
    "pytest-playwright>=0.6.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
research-platform = "research_platform:main"
research-cli = "src.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501", "B008", "C901", "RUF012"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
